refactor copy of tiles db and cleanup
